Added Math

+++ db/migrate/20170220192432_create_restaurants.rb
@@ -4,7 +4,7 @@ class CreateRestaurants < ActiveRecord::Migration[5.0]
       t.string :name
       t.string :logo_url
       t.string :menu_link
-      t.integer :delivery_fee
+      t.float :delivery_fee
     end
   end
 end
