+++ db/migrate/20170220192439_create_orders.rb
@@ -2,8 +2,8 @@ class CreateOrders < ActiveRecord::Migration[5.0]
   def change
     create_table :orders do |t|
       t.string :head_patron
-      t.integer :total_price
-      t.integer :tip
+      t.float :total_price
+      t.float :tip
       t.belongs_to :restaurant
     end
   end
