+++ db/migrate/20170220192446_create_patrons.rb
@@ -3,7 +3,7 @@ class CreatePatrons < ActiveRecord::Migration[5.0]
     create_table :patrons do |t|
       t.string :name
       t.string :personal_order
-      t.integer :personal_price
+      t.float :personal_price
       t.belongs_to :order
     end
   end
