+++ db/migrate/20170221012827_create_items.rb
@@ -2,9 +2,9 @@ class CreateItems < ActiveRecord::Migration[5.0]
   def change
     create_table :items do |t|
       t.string :name
-      t.integer :price
+      t.float :price
       t.string :modifications
-      t.integer :modifications_price
+      t.float :modifications_price
       t.belongs_to :patron
     end
   end
